fix: Fix Edit user information function
