+++ js/requests.js
@@ -1,5 +1,11 @@
 import { setItemWithExpireTime } from "./signup.js";
-import { loginBtnEl, idboxEl, pwboxEl, loginErrorBox } from "./main.js";
+import {
+  loginBtnEl,
+  idboxEl,
+  pwboxEl,
+  loginErrorBox,
+  userInfoName,
+} from "./main.js";
 
 const API_KEY = `FcKdtJs202209`;
 const USER_NAME = `imyeji`;
@@ -81,7 +87,6 @@ export async function logout() {
   loginBtnEl.textContent = "로그인/가입";
 }
 
-export const userInfoName = document.getElementById("user-info-name");
 // 인증 확인 api
 export async function authLogin() {
   const tokenValue = localStorage.getItem("token");
@@ -110,7 +115,7 @@ export async function authLogin() {
 }
 
 // 사용자 정보 수정 api
-async function editUser(displayName, oldPassword, newPassword) {
+export async function editUser(displayName, oldPassword, newPassword) {
   const tokenValue = localStorage.getItem("token");
   const token = JSON.parse(tokenValue).value;
   const res = await fetch(
@@ -133,17 +138,3 @@ async function editUser(displayName, oldPassword, newPassword) {
   const json = await res.json();
   console.log("Response:", json);
 }
-
-// 이름 옆에 변경 버튼 누르면 이름 변경되도록 만들기
-const nameChangeBtn = document.querySelector(".name-change-btn");
-nameChangeBtn.addEventListener("click", async () => {
-  await editUser(userInfoName.value);
-});
-
-// 비밀번호 변경 버튼 누르면 비밀번호 변경되도록 만들기
-const userInfoPw = document.getElementById("user-info-pwd");
-const userInfoNewPw = document.getElementById("user-info-new-pwd");
-const pwChangeBtn = document.querySelector(".pw-change-btn");
-pwChangeBtn.addEventListener("click", async () => {
-  await editUser(userInfoName.value, userInfoPw.value, userInfoNewPw.value);
-});
